requirements.txt
web3>=6.0.0
requests>=2.31.0
pandas>=2.0.0
scikit-learn>=1.3.0
torch>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
websocket-client>=1.6.0
ccxt>=4.0.0
